Move stdlib into shared folder
