+++ libs/@blockprotocol/graph/src/shared/stdlib/bound.ts
@@ -1,7 +1,4 @@
-import {
-  TemporalBound,
-  TimeInterval,
-} from "../shared/types/temporal-versioning.js";
+import { TemporalBound, TimeInterval } from "../types/temporal-versioning.js";
 
 export const compareBounds = (
   left: TemporalBound,
