+++ libs/@blockprotocol/graph/src/shared/stdlib/interval.ts
@@ -4,7 +4,7 @@ import {
   TemporalBound,
   TimeInterval,
   Timestamp,
-} from "../shared/types/temporal-versioning.js";
+} from "../types/temporal-versioning.js";
 import { boundIsAdjacentToBound, compareBounds } from "./bound.js";
 
 /**
