+++ libs/@blockprotocol/graph/src/shared/stdlib/must-be-defined.ts
