+++ libs/@blockprotocol/graph/src/shared/stdlib/subgraph/element/data-type.ts
@@ -5,13 +5,10 @@ import {
   VersionedUri,
 } from "@blockprotocol/type-system/slim";
 
-import { typedValues } from "../../../shared.js";
-import { DataTypeWithMetadata } from "../../../shared/types/ontology/data-type.js";
-import {
-  OntologyTypeVertexId,
-  Subgraph,
-} from "../../../shared/types/subgraph.js";
-import { isDataTypeVertex } from "../../../shared/types/subgraph/vertices.js";
+import { typedValues } from "../../../../shared.js";
+import { DataTypeWithMetadata } from "../../../types/ontology/data-type.js";
+import { OntologyTypeVertexId, Subgraph } from "../../../types/subgraph.js";
+import { isDataTypeVertex } from "../../../types/subgraph/vertices.js";
 
 /**
  * Returns all `DataTypeWithMetadata`s within the vertices of the subgraph
