+++ libs/@blockprotocol/graph/src/stdlib.ts
@@ -1,7 +1,7 @@
 /**
  * The extended standard library of functions for interacting with a {@link Subgraph}.
  */
-export { compareBounds } from "./stdlib/bound.js";
+export { compareBounds } from "./shared/stdlib/bound.js";
 export {
   intervalCompareWithInterval,
   intervalContainsInterval,
@@ -16,45 +16,45 @@ export {
   intervalUnionWithInterval,
   sortIntervals,
   unionOfIntervals,
-} from "./stdlib/interval.js";
-export { buildSubgraph } from "./stdlib/subgraph/builder.js";
-export { getPropertyTypesReferencedByEntityType } from "./stdlib/subgraph/edge/entity-type.js";
+} from "./shared/stdlib/interval.js";
+export { buildSubgraph } from "./shared/stdlib/subgraph/builder.js";
+export { getPropertyTypesReferencedByEntityType } from "./shared/stdlib/subgraph/edge/entity-type.js";
 export {
   getIncomingLinksForEntity,
   getLeftEntityForLinkEntity,
   getOutgoingLinkAndTargetEntities,
   getOutgoingLinksForEntity,
   getRightEntityForLinkEntity,
-} from "./stdlib/subgraph/edge/link-entity.js";
+} from "./shared/stdlib/subgraph/edge/link-entity.js";
 export {
   getDataTypeById,
   getDataTypeByVertexId,
   getDataTypes,
   getDataTypesByBaseUri,
-} from "./stdlib/subgraph/element/data-type.js";
+} from "./shared/stdlib/subgraph/element/data-type.js";
 export {
   getEntities,
   getEntityRevision,
   getEntityRevisionsByEntityId,
-} from "./stdlib/subgraph/element/entity.js";
+} from "./shared/stdlib/subgraph/element/entity.js";
 export {
   getEntityTypeById,
   getEntityTypeByVertexId,
   getEntityTypes,
   getEntityTypesByBaseUri,
-} from "./stdlib/subgraph/element/entity-type.js";
-export { mapElementsIntoRevisions } from "./stdlib/subgraph/element/map-revisions.js";
+} from "./shared/stdlib/subgraph/element/entity-type.js";
+export { mapElementsIntoRevisions } from "./shared/stdlib/subgraph/element/map-revisions.js";
 export {
   getPropertyTypeById,
   getPropertyTypeByVertexId,
   getPropertyTypes,
   getPropertyTypesByBaseUri,
-} from "./stdlib/subgraph/element/property-type.js";
+} from "./shared/stdlib/subgraph/element/property-type.js";
 export {
   getRoots,
   isDataTypeRootedSubgraph,
   isEntityRootedSubgraph,
   isEntityTypeRootedSubgraph,
   isPropertyTypeRootedSubgraph,
-} from "./stdlib/subgraph/roots.js";
-export { getLatestInstantIntervalForSubgraph } from "./stdlib/subgraph/temporal-axes.js";
+} from "./shared/stdlib/subgraph/roots.js";
+export { getLatestInstantIntervalForSubgraph } from "./shared/stdlib/subgraph/temporal-axes.js";
